Added more stemmer tests, refactored and redistributed old tests related to text pre-processing.
